+++ sdk/appconfiguration/azure-data-appconfiguration/src/samples/java/com/azure/data/appconfiguration/ReadmeSamples.java
@@ -21,6 +21,7 @@ import com.azure.data.appconfiguration.models.ConfigurationSettingsFilter;
 import com.azure.data.appconfiguration.models.ConfigurationSnapshot;
 import com.azure.data.appconfiguration.models.FeatureFlagConfigurationSetting;
 import com.azure.data.appconfiguration.models.FeatureFlagFilter;
+import com.azure.data.appconfiguration.models.LabelSelector;
 import com.azure.data.appconfiguration.models.SecretReferenceConfigurationSetting;
 import com.azure.data.appconfiguration.models.SettingSelector;
 import com.azure.data.appconfiguration.models.SnapshotSelector;
@@ -466,6 +467,16 @@ public class ReadmeSamples {
         }
         // END: readme-sample-listSettingsInSnapshot
 
+    }
+
+    public void listLabels() {
+        // BEGIN: readme-sample-listLabels
+        String labelFilter = "{labelNamePrefix}*";
+        configurationClient.listLabels(new LabelSelector().setLabelFilter(labelFilter))
+                .forEach(label -> {
+                    System.out.println("label name = " + label);
+                });
+        // END: readme-sample-listLabels
     }
 
     private void updateConfiguration(ConfigurationSetting setting) {