+++ sdk/appconfiguration/azure-data-appconfiguration/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- Added a new method `listLabels` to support listing labels capabilities.
+
 ### Breaking Changes
 
 ### Bugs Fixed